pyproject.toml
[project]
name = "daishan-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.1",
    "fastapi==0.104.1",
    "openai>=2.15.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.6",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "uvicorn[standard]==0.24.0",
]
